rename core.py -> types.py & remove dask dep
